+++ R/variousOtherHelperFunctions.R
@@ -2491,6 +2491,8 @@ lapplyLeaf <- function(x, f) {
 # reformat matrix for display ----
 showMatrixCorner <- function(x, nRows = 4, nCols = 3, digits = 3) {
   
+  x <- stripClass(x)
+  
   stopifnot(is.matrix(x))
   
   nr <- nrow(x)
@@ -2539,7 +2541,8 @@ showMatrixCorner <- function(x, nRows = 4, nCols = 3, digits = 3) {
 ## color to mimic pillar output (not a function)
 crayon_grey_0.6 <- crayon::make_style(grDevices::grey(0.6), grey = TRUE) # mimic pillar
 
-# function to strip camtrapR class and associated camtrapR attributes (for testing only)
+
+# function to strip camtrapR class and associated camtrapR attributes
 stripClass <- function(x) {
   class(x) <- setdiff(class(x), c("camOp", "records", "detHist"))
   all_attributes <- names(attributes(x))